更新数据处理逻辑

+++ Core/Librarys/SQLite/SQLiteBuilder.cs
@@ -110,7 +110,7 @@ namespace Core.Librarys.SQLite
             }
 
             IsSelfChecking = false;
-            
+
             return true;
         }
 
@@ -121,6 +121,8 @@ namespace Core.Librarys.SQLite
         {
             try
             {
+                bool isNew = true;
+
                 //  处理前先复制一份数据库文件以防万一
                 string dir = Path.Combine(Path.GetDirectoryName(dbFile), "backup");
                 string backupName = Path.Combine(dir, $"data.handle.backup");
@@ -132,8 +134,13 @@ namespace Core.Librarys.SQLite
                 {
                     File.Delete(backupName);
                 }
-                File.Copy(dbFile, backupName);
 
+                if (File.Exists(dbFile))
+                {
+                    isNew = false;
+
+                    File.Copy(dbFile, backupName);
+                }
                 var modelInfos = GetModelInfos();
 
                 var modelInfosDb = GetModelInfosForDb();
@@ -170,6 +177,7 @@ namespace Core.Librarys.SQLite
                     }
 
                 }
+
 
                 HandleVersion1002Migrate();
